Split into two classes
